+++ src/blockly/constants/index.js
@@ -1,8 +1,0 @@
-/**
- * @author Luciano Graziani @lgraziani2712
- * @license {@link http://www.opensource.org/licenses/mit-license.php|MIT License}
- *
- * @flow
- */
-export * as colors from './colors';
-export * as blockNames from './blockNames';