+++ src/blockly/localize/es.js
@@ -10,10 +10,13 @@ Blockly.Msg.ACTION_CONTAINER_TITLE = 'Poné las piezas acá adentro 😁';
 Blockly.Msg.ACTION_CONTAINER_TOOLTIP = 'Es el bloque principal. Aquí es donde se deben poner todas las acciones';
 
 Blockly.Msg.MOVE_LEFT_TITLE = '⬅  Mover a la izquierda';
-Blockly.Msg.MOVE_LEFT_TOOLTIP = 'Mueve el personaje un paso a la izquierda';
+Blockly.Msg.MOVE_LEFT_TOOLTIP = 'Mueve el número un paso a la izquierda';
 Blockly.Msg.MOVE_RIGHT_TITLE = '➡  Mover a la derecha';
-Blockly.Msg.MOVE_RIGHT_TOOLTIP = 'Mueve el personaje un paso a la derecha';
+Blockly.Msg.MOVE_RIGHT_TOOLTIP = 'Mueve el número un paso a la derecha';
 Blockly.Msg.MOVE_FORWARD_TITLE = '⬆  Mover hacia delante';
-Blockly.Msg.MOVE_FORWARD_TOOLTIP = 'Mueve el personaje un paso adelante';
+Blockly.Msg.MOVE_FORWARD_TOOLTIP = 'Mueve el número un paso adelante';
 Blockly.Msg.MOVE_BACKWARD_TITLE = '⬇  Mover hacia atrás';
-Blockly.Msg.MOVE_BACKWARD_TOOLTIP = 'Mueve el personaje un paso atrás';
+Blockly.Msg.MOVE_BACKWARD_TOOLTIP = 'Mueve el número un paso atrás';
+
+Blockly.Msg.LEAVE_MAZE_TITLE = '🚪🏃 Salir del laberinto';
+Blockly.Msg.LEAVE_MAZE_TOOLTIP = 'Si el número está sobre una salida, lo mueve hacia la línea numérica';