+++ src/constants/actions.js
@@ -6,8 +6,9 @@
  */
 
 export const ACTION_CONTAINER = 'action_container';
+export const LEAVE_MAZE = 'leave_maze';
 
-export const MOVE_BACKWARD = 'move_backward';
 export const MOVE_FORWARD = 'move_forward';
-export const MOVE_LEFT = 'move_left';
 export const MOVE_RIGHT = 'move_right';
+export const MOVE_BACKWARD = 'move_backward';
+export const MOVE_LEFT = 'move_left';