+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## HEAD
 
+- 🐛 [2017-05-18] `engine/containers/mazeEngineGenerator`: update to v3.4.2.
+  - Fix #12. Definitely! But with a better approach: it requires the new `leave_maze` action to be called for the number to leave.
+- 🚀 [2017-05-18] `engine/errors`: update to v1.1.0.
+  - 🚀 Add `MazeStarvationError`: thrown when the number did not leave.
+  - 🚀 Add `MazePathOverflow`: thrown when the number tries to move forward at the end of the maze.
+- 🚀 [2017-05-18] `blockly/components`: update to v1.1.0.
+  - 🚀 Add `LeaveMaze`.
+  - ⚙ Refactor `index.js` to load every blockly component automatically.
+- 🐛 [2017-05-18] `engine/containers/mazeEngineGenerator`: update to v3.4.1.
+  - Fix #12. Now checks not only if the number is at one exit but if the next action moves the number forward.
+- ⚙ [2017-05-18] `constants`: every group of constants are only in `src/constants`.
 - 🚀 [2017-05-12] `engine/containers/mazeEngineGenerator`: update to v3.4.0.
   - 🚀 Update methods to match the new `beSad` number's animation.
 - 🚀 [2017-05-12] `engine/components/numericLineGenerator`: update to v3.2.0.